+++ app/scripts/Track.js
@@ -180,6 +180,14 @@ class Track {
   respondsToPosition(x, y) {
     return this.isWithin(x, y);
   }
+
+  zoomedY(trackY, kMultiplier) {
+
+  }
+
+  movedY(dY) {
+
+  }
 }
 
 export default Track;